backend/requirements.txt
# Django Core
Django==4.2.7
djangorestframework==3.14.0

# Database - SQL Server
mssql-django==1.3
pyodbc==5.0.1

# Filters
django-filter==23.3

# CORS
django-cors-headers==4.3.0

# Timezone handling
pytz==2023.3

# Excel/CSV export
openpyxl==3.1.2

# Environment variables
python-dotenv==1.0.0

# Production server
gunicorn==21.2.0

# Development tools
django-debug-toolbar==4.2.0
